requirements.txt
streamlit
sentence-transformers
scikit-learn
PyPDF2
openai
numpy
faiss-cpu
